feat: Release calculator job with total production quantity (#265)
* Remove current calculator job
* Release new calculator job
