+++ source/databricks/package/__init__.py
@@ -15,4 +15,3 @@
 from .spark_initializor import initialize_spark
 from .integration_events_persister import integration_events_persister
 from .balance_fixing_total_production import calculate_balance_fixing_total_production
-from .calculator import calculator
